A script used to list screensaver inhibitors
This script can be used to list the screensaver inhibitors. Those processes that are currently running and prevent your screensaver from kicking in. When I go to bed, I like to lock my screen, only in the morning at times it's on. That's becomes a process (often Firefox) has an inhibitor in effect. The idea of the script is to find out which that is and close that tab or tool for the night.